# docs/swagger.yaml
basePath: /v2
definitions:
...
  view.ErrorResult:
    properties:
      error:
        type: string
    type: object
...
  view.WorkflowResponse:
    properties:
      mode:
        type: integer
      stages:
        type: string
      status:
        type: integer
      step:
        type: integer
      type:
        type: integer
      workflow_id:
        type: string
    type: object
host: 127.0.0.1:2002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: v1.8 存储管理系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: PolarDB Storage Manage System
  version: "1.0"
paths:
...
  /workflows/:workflowId:
    get:
      consumes:
      - application/json
      description: 用于查询 WorkflowEntity
      parameters:
      - description: 请求参数
        in: query
        name: workflowId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 查询 WorkflowEntity
      tags:
      - Workflow 管理
swagger: "2.0"